+++ Source/EventFlow.PostgreSql/EventStores/PostgresSqlEventPersistence.cs
@@ -69,24 +69,24 @@ namespace EventFlow.PostgreSql.EventStores
             var startPosition = globalPosition.IsStart
                 ? 0
                 : long.Parse(globalPosition.Value);
-            var endPosition = startPosition + pageSize;
 
             const string sql = @"
                 SELECT
                     GlobalSequenceNumber, BatchId, AggregateId, AggregateName, Data, Metadata, AggregateSequenceNumber
                 FROM EventFlow
                 WHERE
-                    GlobalSequenceNumber >= @FromId AND GlobalSequenceNumber <= @ToId
+                    GlobalSequenceNumber >= @startPosition
                 ORDER BY
-                    GlobalSequenceNumber ASC;";
+                    GlobalSequenceNumber ASC
+                LIMIT @pageSize;";
             var eventDataModels = await _connection.QueryAsync<EventDataModel>(
-                Label.Named("postgresql-fetch-events"),
-                cancellationToken,
-                sql,
-                new
+                    Label.Named("postgresql-fetch-events"),
+                    cancellationToken,
+                    sql,
+                    new
                     {
-                        FromId = startPosition,
-                        ToId = endPosition,
+                        startPosition,
+                        pageSize
                     })
                 .ConfigureAwait(false);